Create sample1.cob
Simply displaying "Hello World" is not interesting, so I tried to display it again if "0" is entered after it is displayed.

The first line contains the number of columns. If the 7th digit is "*", it becomes a comment line. Normally, digits 1 to 6 are line numbers. It doesn't have to be sorted. Although it is not limited to COBOL, in the past, sources and data were created with paper punch cards, and the order of the paper punch cards was accidentally changed. For that reason, we added a sequence number and added a process to sort the paper punch cards when they were inserted.