FortuneTelling
[hiroppy123@fedora ~]$ javac FortuneTelling.java
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 1
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 3
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 3
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 3
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 2
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 6
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 4
Is it Daikichi? false
[hiroppy123@fedora ~]$ java FortuneTelling
The extracted number is: 7
Is it Daikichi? true