fortuneballdaikiti
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 9
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 6
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 9
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 2
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 1
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 5
Is it Daikichi? No
[hiroppy123@fedora ~]$ python3 daikiti.py
The extracted number is: 7
Is it Daikichi? Yes